[bugfix] Ensure all contract parameters have been provided
In the "finalize" handler, there is no check to ensure all the
parameters have been sent to the plugin. A user could then sign
incomplete contract data, or uninitialized data could be displyed
on the screen.

This fix checks if no more data has to be received by the plugin.